backend/requirements.txt
fastapi
uvicorn
sqlalchemy
mariadb
python-dotenv
fastapi-users
pydantic
